Changed command groups to be a max size.
Since we hit a UDP max packet size we needed to split the data we send
into multiple calls.

+++ ChristmasLightController/src/ca/cmfly/controller/LightId.java
@@ -1,10 +1,10 @@
 package ca.cmfly.controller;
 
 public class LightId {
-	public int strandNum;
-	public int lightNum;
+	public byte strandNum;
+	public byte lightNum;
 
-	public LightId(int strandNum, int lightNum) {
+	public LightId(byte strandNum, byte lightNum) {
 		this.strandNum = strandNum;
 		this.lightNum = lightNum;
 	}
